+++ src/LinkMobility/ILinkMobilityClient.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using AMWD.Net.Api.LinkMobility.Requests;
 
 namespace AMWD.Net.Api.LinkMobility
 {
@@ -14,6 +13,13 @@ namespace AMWD.Net.Api.LinkMobility
 		/// </summary>
 		/// <param name="request">The request data.</param>
 		/// <param name="cancellationToken">A cancellation token to propagate notification that operations should be canceled.</param>
-		Task<SendTextMessageResponse> SendTextMessage(SendTextMessageRequest request, CancellationToken cancellationToken = default);
+		Task<SendMessageResponse> SendTextMessage(SendTextMessageRequest request, CancellationToken cancellationToken = default);
+
+		/// <summary>
+		/// Sends a binary message to a list of recipients.
+		/// </summary>
+		/// <param name="request">The request data.</param>
+		/// <param name="cancellationToken">A cancellation token to propagate notification that operations should be canceled.</param>
+		Task<SendMessageResponse> SendBinaryMessage(SendBinaryMessageRequest request, CancellationToken cancellationToken = default);
 	}
 }